Upgrade Next.js to 14.x - Also update next-auth to beta versions that support Next 14.x
