+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,53 +1,1 @@
-import { PrismaAdapter } from "@next-auth/prisma-adapter";
-import NextAuth, { AuthOptions } from "next-auth";
-import GithubProvider from "next-auth/providers/github";
-import { prisma } from "@/lib/prisma";
-
-export const nextAuthOptions: AuthOptions = {
-  adapter: PrismaAdapter(prisma),
-  callbacks: {
-    async jwt({ token, user }) {
-      const dbUser = await prisma.user.findFirst({
-        where: {
-          email: token.email,
-        },
-      });
-
-      if (!dbUser) {
-        if (user) {
-          token.id = user.id;
-        }
-        return token;
-      }
-
-      return {
-        id: dbUser.id,
-        name: dbUser.name,
-        email: dbUser.email,
-        picture: dbUser.image,
-      };
-    },
-    async session({ session, token }) {
-      if (token) {
-        session.user.id = token.id;
-        session.user.name = token.name;
-        session.user.email = token.email;
-        session.user.image = token.picture;
-      }
-
-      return session;
-    },
-  },
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_CLIENT_ID || "",
-      clientSecret: process.env.GITHUB_CLIENT_SECRET || "",
-    }),
-  ],
-  secret: process.env.NEXTAUTH_SECRET || "",
-  session: {
-    strategy: "jwt",
-  },
-};
-const handler = NextAuth(nextAuthOptions);
-export { handler as GET, handler as POST };
+export { GET, POST } from "@/lib/auth";
