+++ src/app/page.tsx
@@ -1,14 +1,15 @@
-import { getServerSession } from "next-auth";
-import { nextAuthOptions } from "./api/auth/[...nextauth]/route";
-
-async function getSession() {
-  return await getServerSession(nextAuthOptions);
-}
-
-async function getCurrentUser() {
-  const session = await getSession();
-  return session?.user;
-}
+import { getCurrentUser } from "@/lib/session";
+// import { getServerSession } from "next-auth";
+// import { nextAuthOptions } from "./api/auth/[...nextauth]/route";
+//
+// async function getSession() {
+//   return await getServerSession(nextAuthOptions);
+// }
+//
+// async function getCurrentUser() {
+//   const session = await getSession();
+//   return session?.user;
+// }
 
 export default async function Home() {
   const user = await getCurrentUser();
